Chore: 장바구니 코드 수정

+++ src/main/java/com/palleteforco/palleteforco/domain/cart/dto/CartDto.java
@@ -19,5 +19,5 @@ public class CartDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime cart_date;
-    private Long cart_status;
+    private int cart_status;
 }
